backend-v2/requirements.txt
Django==2.2.1
djangorestframework==3.9.4
mysql-connector-python==8.0.16